# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id: string;
  title: string;
  thumbnail?: string;
}

export interface TelegramLoginInput {
  telegramId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface UserProfile {
  id: string;
  telegramId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isNewUser?: boolean;
}

export interface HistoryInput {
  telegramId: string;
  song: Song;
}

export interface HistoryResponse {
  history: Song[];
}

export interface FavoriteInput {
  telegramId: string;
  songId: string;
}

export interface FavoritesResponse {
  favorites: string[];
}

export interface PlaylistInput {
  telegramId: string;
  song: Song;
}

export interface PlaylistResponse {
  playlist: Song[];
}

export interface SearchResult {
  id: string;
  title: string;
  thumbnail: string;
  channelTitle?: string;
}

export interface SearchResults {
  items: SearchResult[];
}

export type SearchMusicParams = {
/**
 * Search query
 */
q: string;
};