# lib/api-zod/src/generated/types/favoriteInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteInput {
  telegramId: string;
  songId: string;
}