# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface PlaylistInput {
  telegramId: string;
  song: Song;
}