# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  id: string;
  title: string;
  thumbnail: string;
  channelTitle?: string;
}