# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from './searchResult';

export interface SearchResults {
  items: SearchResult[];
}