# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RADHA MUSIC API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  telegramId: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isNewUser?: boolean;
}